Update use-device-guard-with-configuration-manager.md
Please change the title to not let customer's skip the before you begin section if they are looking into group policy deployment method. We had a current situation where customer did not read the Before you begin section which had mentioned that hta application have an exception. This was because the title says for Config Manager and they were using Group Policy method.

Also please use the name of the navigation title as the heading of the article.